+++ src/registration-columns/RegistrationColumns.js
