+++ src/registration-row/RegistrationRow.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import {TextInput, CheckBox, Dropdown} from '../registration-columns';
+import {TextInput, CheckBox, Dropdown} from '../registration-columns/RegistrationColumns';
 
 import {TEXT, CHECKBOX, DROPDOWN} from '../registration-columns/types';
